+++ tests/clv/test_plotting.py
@@ -14,11 +14,6 @@ from pymc_marketing.clv.plotting import (
 )
 
 
-@pytest.fixture(scope="module")
-def test_summary_data() -> pd.DataFrame:
-    return pd.read_csv("tests/clv/datasets/test_summary_data.csv", index_col=0)
-
-
 @pytest.mark.parametrize(
     "kwargs",
     [
